+++ SteelBot/DataProviders/SubProviders/SelfRoleProvider.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.Logging;
 using SteelBot.Database;
 using SteelBot.Database.Models;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,8 +12,8 @@ namespace SteelBot.DataProviders.SubProviders;
 
 public class SelfRolesProvider
 {
-    private readonly ILogger<SelfRolesProvider> _logger;
     private readonly IDbContextFactory<SteelBotContext> _dbContextFactory;
+    private readonly ILogger<SelfRolesProvider> _logger;
 
     private readonly ConcurrentDictionary<ulong, ConcurrentDictionary<ulong, SelfRole>> _selfRolesByGuildAndId;
 
@@ -35,10 +36,7 @@ public class SelfRolesProvider
             allRoles = db.SelfRoles.AsNoTracking().Include(sr => sr.Guild).ToArray();
         }
 
-        foreach (var role in allRoles)
-        {
-            AddRoleToInternalCache(role.Guild.DiscordId, role);
-        }
+        foreach (var role in allRoles) AddRoleToInternalCache(role.Guild.DiscordId, role);
     }
 
     private void AddRoleToInternalCache(ulong guildId, SelfRole role)
@@ -49,23 +47,14 @@ public class SelfRolesProvider
 
     private void RemoveRoleFromInternalCache(ulong guildId, ulong roleId)
     {
-        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles))
-        {
-            roles.TryRemove(roleId, out _);
-        }
+        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles)) roles.TryRemove(roleId, out _);
     }
 
-    public bool BotKnowsRole(ulong guildId, ulong roleId)
-    {
-        return _selfRolesByGuildAndId.TryGetValue(guildId, out var roles) ? roles.ContainsKey(roleId) : false;
-    }
+    public bool BotKnowsRole(ulong guildId, ulong roleId) => _selfRolesByGuildAndId.TryGetValue(guildId, out var roles) ? roles.ContainsKey(roleId) : false;
 
     public bool TryGetRole(ulong guildId, ulong roleId, out SelfRole role)
     {
-        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles))
-        {
-            return roles.TryGetValue(roleId, out role);
-        }
+        if (_selfRolesByGuildAndId.TryGetValue(guildId, out var roles)) return roles.TryGetValue(roleId, out role);
         role = null;
         return false;
     }
@@ -73,7 +62,7 @@ public class SelfRolesProvider
     public bool TryGetRole(ulong guildId, string roleName, out SelfRole role)
     {
         role = _selfRolesByGuildAndId.TryGetValue(guildId, out var guildRoles)
-            ? guildRoles.Values.FirstOrDefault(x => x.RoleName.Equals(roleName, System.StringComparison.OrdinalIgnoreCase))
+            ? guildRoles.Values.FirstOrDefault(x => x.RoleName.Equals(roleName, StringComparison.OrdinalIgnoreCase))
             : null;
         return role != null;
     }
@@ -85,24 +74,19 @@ public class SelfRolesProvider
             roles = indexedRoles.Values.ToList();
             return true;
         }
+
         roles = new List<SelfRole>();
         return false;
     }
 
     public async Task AddRole(ulong guildId, SelfRole role)
     {
-        if (!BotKnowsRole(guildId, role.DiscordRoleId))
-        {
-            await InsertSelfRole(guildId, role);
-        }
+        if (!BotKnowsRole(guildId, role.DiscordRoleId)) await InsertSelfRole(guildId, role);
     }
 
     public async Task RemoveRole(ulong guildId, ulong roleId)
     {
-        if (TryGetRole(guildId, roleId, out var role))
-        {
-            await DeleteSelfRole(guildId, role);
-        }
+        if (TryGetRole(guildId, roleId, out var role)) await DeleteSelfRole(guildId, role);
     }
 
     private async Task InsertSelfRole(ulong guildId, SelfRole role)
@@ -117,13 +101,9 @@ public class SelfRolesProvider
         }
 
         if (writtenCount > 0)
-        {
             AddRoleToInternalCache(guildId, role);
-        }
         else
-        {
             _logger.LogError("Writing Self Role {RoleName} for Guild {GuildId} to the database inserted no entities. The internal cache was not changed", role.RoleName, guildId);
-        }
     }
 
     private async Task DeleteSelfRole(ulong guildId, SelfRole role)
@@ -138,12 +118,8 @@ public class SelfRolesProvider
         }
 
         if (writtenCount > 0)
-        {
             RemoveRoleFromInternalCache(guildId, role.DiscordRoleId);
-        }
         else
-        {
             _logger.LogWarning("Deleting Self Role {RoleName} for Guild {GuildId} from the database deleted no entities. The internal cache was not changed", role.RoleName, guildId);
-        }
     }
 }
